--- src/main/java/com/ll/playon/domain/member/entity/Gender.java
@@ -1,5 +1,0 @@
-package com.ll.playon.domain.member.entity;
-
-public enum Gender {
-    MALE, FEMALE
-}
