--- src/main/java/com/ll/playon/domain/member/entity/Role.java
@@ -1,6 +1,0 @@
-package com.ll.playon.domain.member.entity;
-
-public enum Role {
-    ADMIN,
-    USER
-}
