+++ src/main/java/com/ll/playon/domain/member/entity/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.ll.playon.domain.member.entity.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
