+++ src/main/java/com/ll/playon/domain/member/entity/enums/PlayStyle.java
@@ -1,4 +1,4 @@
-package com.ll.playon.domain.member.entity;
+package com.ll.playon.domain.member.entity.enums;
 
 public enum PlayStyle {
     CASUAL, HARDCORE, SPEEDRUN, COMPLETIONIST
