+++ src/main/java/com/ll/playon/domain/member/entity/enums/PreferredGenres.java
@@ -1,4 +1,4 @@
-package com.ll.playon.domain.member.entity;
+package com.ll.playon.domain.member.entity.enums;
 
 public enum PreferredGenres {
     FPS, RPG, MOBA, SPORTS, PUZZLE
