+++ src/main/java/com/ll/playon/domain/member/entity/enums/Role.java
@@ -1,0 +1,6 @@
+package com.ll.playon.domain.member.entity.enums;
+
+public enum Role {
+    ADMIN,
+    USER
+}
