+++ src/main/java/com/ll/playon/global/initData/BaseInitData.java
@@ -6,9 +6,9 @@ import com.ll.playon.domain.guild.guild.repository.GuildRepository;
 import com.ll.playon.domain.guild.guildMember.entity.GuildMember;
 import com.ll.playon.domain.guild.guildMember.enums.GuildRole;
 import com.ll.playon.domain.guild.guildMember.repository.GuildMemberRepository;
-import com.ll.playon.domain.member.MemberRepository;
-import com.ll.playon.domain.member.MemberService;
-import com.ll.playon.domain.member.MemberSteamDataRepository;
+import com.ll.playon.domain.member.repository.MemberRepository;
+import com.ll.playon.domain.member.service.MemberService;
+import com.ll.playon.domain.member.repository.MemberSteamDataRepository;
 import com.ll.playon.domain.member.entity.Member;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
