+++ src/main/java/com/ll/playon/global/security/UserContext.java
@@ -1,6 +1,6 @@
 package com.ll.playon.global.security;
 
-import com.ll.playon.domain.member.MemberService;
+import com.ll.playon.domain.member.service.MemberService;
 import com.ll.playon.domain.member.entity.Member;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -104,11 +104,31 @@ public class UserContext {
 
     public Member getActor() {
         // TODO : 실제 인증된 사용자 조회로 바꾸기
-        return memberService.findById(1L);
+        return memberService.findById(1L).get();
     }
 
-    public Optional<Member> findActor() {
-        // TODO : 실제 로직으로 수정
-        return Optional.empty();
-    }
+//    // 요청을 보낸 사용자의 인증 정보를 가져와 해당 사용자를 조회, 시큐리티 내부에 인증된 사용자를 반환
+//    public Member getActor() {
+//        return Optional.ofNullable(
+//                        SecurityContextHolder
+//                                .getContext()
+//                                .getAuthentication()
+//                )
+//                .map(Authentication::getPrincipal)
+//                .filter(principal -> principal instanceof SecurityUser)
+//                .map(principal -> (SecurityUser) principal)
+//                .map(securityUser -> new Member(securityUser.getId(), securityUser.getUsername()))
+//                .orElse(null);
+//    }
+//
+//    // 요청을 보낸 사용자의 인증 정보를 가져와 실제 DB에 저장된 user 찾기
+//    public Optional<Member> findActor() {
+//        Member actor = getActor();
+//
+//        if (actor == null) {
+//            return Optional.empty();
+//        }
+//
+//        return memberService.findById(actor.getId());
+//    }
 }
